requirements.txt
# API
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1

# HTTP Client (worker -> ACE-Step API)
httpx==0.28.1

# System monitoring
psutil>=5.9.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
pyyaml==6.0.2

# Linting
ruff==0.9.4
